+++ R/classes.R
@@ -34,7 +34,7 @@ check_classes <- function(x, classes = character(0),
   x_classes <- class(x)
   
   if(!length(classes) && exclusive) 
-    error("all objects must have a class (define classes or set exclusive = FALSE)")
+    err("all objects must have a class (define classes or set exclusive = FALSE)")
   
   if (identical(length(classes), 1L) && !exclusive) {
     check_inherits(x, classes, x_name = x_name, error = error)
@@ -44,19 +44,19 @@ check_classes <- function(x, classes = character(0),
   if (exclusive) {
     if (order) {
       if (!identical(x_classes, classes))
-        on_fail(x_name, " classes must be identical to ", punctuate(classes, "and"), error = error)
+        chk_fail(x_name, " classes must be identical to ", cc(classes, "and"), error = error)
     } else {
       if (!identical(sort(x_classes), sort(classes)))
-        on_fail(x_name, " classes must include and only include ", punctuate(classes, "and"), error = error)
+        chk_fail(x_name, " classes must include and only include ", cc(classes, "and"), error = error)
     }
   } else {
     x_classes <- x_classes[x_classes %in% classes]
     if (order) {
       if (!identical(x_classes, classes))
-        on_fail(x_name, " classes must include ", punctuate(classes, "and"), " in that order", error = error)
+        chk_fail(x_name, " classes must include ", cc(classes, "and"), " in that order", error = error)
     } else {
       if (!identical(sort(x_classes), sort(classes)))
-        on_fail(x_name, " classes must include ", punctuate(classes, "and"), error = error)
+        chk_fail(x_name, " classes must include ", cc(classes, "and"), error = error)
     }
   }
   invisible(x)