+++ R/levels.R
@@ -22,31 +22,31 @@ check_levels <- function(x, levels, exclusive = TRUE, order = TRUE,
                          error = TRUE) {
   x_name <- deparse_x_name(x_name)
 
-  check_vector(levels, "", length = c(1L, .Machine$integer.max), unique = FALSE, named = FALSE)
+  check_vector(levels, "", length = c(1L, chk_max_int()), unique = FALSE, named = FALSE)
   check_flag_internal(exclusive)
   check_flag_internal(order)
   check_flag_internal(error)
   
   x_levels <- levels(x)
   
-  if(is.null(x_levels)) on_fail(x_name, " must have a levels attribute", error = error)
+  if(is.null(x_levels)) chk_fail(x_name, " must have a levels attribute", error = error)
 
   if (exclusive) {
     if (order) {
       if (!identical(x_levels, levels))
-        on_fail(x_name, " levels must be identical to ", punctuate(levels, "and"), error = error)
+        chk_fail(x_name, " levels must be identical to ", cc(levels, "and"), error = error)
     } else {
       if (!identical(sort(x_levels), sort(levels)))
-        on_fail(x_name, " levels must include and only include ", punctuate(levels, "and"), error = error)
+        chk_fail(x_name, " levels must include and only include ", cc(levels, "and"), error = error)
     }
   } else {
     x_levels <- x_levels[x_levels %in% levels]
     if (order) {
       if (!identical(x_levels, levels))
-        on_fail(x_name, " levels must include ", punctuate(levels, "and"), " in that order", error = error)
+        chk_fail(x_name, " levels must include ", cc(levels, "and"), " in that order", error = error)
     } else {
       if (!identical(sort(x_levels), sort(levels)))
-        on_fail(x_name, " levels must include ", punctuate(levels, "and"), error = error)
+        chk_fail(x_name, " levels must include ", cc(levels, "and"), error = error)
     }
   }
   invisible(x)